refactor(scraper.aggregator): move to separate pkg
