+++ pkg/scraper/aggregator/aggregator.go
@@ -1,42 +1,27 @@
-package scraper
+package aggregator
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/linden-honey/linden-honey-go/pkg/song"
+
+	"github.com/linden-honey/linden-honey-scraper-go/pkg/scraper"
 )
 
 // Aggregator represents the aggregation service implementation
 type Aggregator struct {
-	services []Service
+	services []scraper.Service
 }
 
 // NewAggregator returns a pointer to the new instance of Aggregator or an error
-func NewAggregator(services ...Service) (*Aggregator, error) {
+func NewAggregator(services ...scraper.Service) (*Aggregator, error) {
 	return &Aggregator{
 		services: services,
 	}, nil
 }
 
-type aggregationErr struct {
-	msg     string
-	reasons []error
-}
-
-func newAggregationErr(msg string, reasons ...error) *aggregationErr {
-	return &aggregationErr{
-		msg:     msg,
-		reasons: reasons,
-	}
-}
-
-func (err *aggregationErr) Error() string {
-	return fmt.Sprintf("%s: %s", err.msg, err.reasons)
-}
-
 // GetSong returns a pointer to the song or an error from aggregated services
-func (a Aggregator) GetSong(ctx context.Context, id string) (*song.Song, error) {
+func (a *Aggregator) GetSong(ctx context.Context, id string) (*song.Song, error) {
 	errs := make([]error, 0)
 	for _, svc := range a.services {
 		s, err := svc.GetSong(ctx, id)
@@ -48,11 +33,11 @@ func (a Aggregator) GetSong(ctx context.Context, id string) (*song.Song, error) 
 		return s, nil
 	}
 
-	return nil, newAggregationErr("failed to scrape a song", errs...)
+	return nil, NewAggregationError("failed to scrape a song", errs...)
 }
 
 // GetSongs returns songs or an error from aggregated services
-func (a Aggregator) GetSongs(ctx context.Context) ([]song.Song, error) {
+func (a *Aggregator) GetSongs(ctx context.Context) ([]song.Song, error) {
 	out := make([]song.Song, 0)
 	errs := make([]error, 0)
 	for _, svc := range a.services {
@@ -66,14 +51,14 @@ func (a Aggregator) GetSongs(ctx context.Context) ([]song.Song, error) {
 	}
 
 	if len(errs) != 0 {
-		return nil, newAggregationErr("failed to aggregate scraped songs", errs...)
+		return nil, NewAggregationError("failed to aggregate scraped songs", errs...)
 	}
 
 	return out, nil
 }
 
 // GetPreviews returns previews or an error from aggregated services
-func (a Aggregator) GetPreviews(ctx context.Context) ([]song.Meta, error) {
+func (a *Aggregator) GetPreviews(ctx context.Context) ([]song.Meta, error) {
 	out := make([]song.Meta, 0)
 	errs := make([]error, 0)
 	for _, svc := range a.services {
@@ -87,7 +72,7 @@ func (a Aggregator) GetPreviews(ctx context.Context) ([]song.Meta, error) {
 	}
 
 	if len(errs) != 0 {
-		return nil, newAggregationErr("failed to aggregate scraped previews", errs...)
+		return nil, NewAggregationError("failed to aggregate scraped previews", errs...)
 	}
 
 	return out, nil
